+++ test-e2e/cypress/support/page-objects/AuthorsLoaded.js
@@ -1,0 +1,12 @@
+import List from "./common/List";
+
+class AuthorsLoaded extends List {
+  constructor() {
+    super({
+      COLUMN: "#authors-loaded-column",
+      CONTAINER: "#authors-loaded-container",
+    });
+  }
+}
+
+export default AuthorsLoaded;